# src/main/resources/api.yml
openapi: 3.1.0
info:
  version: v1
  title: activity-service API
  description: activity-service RESTFul api for openapi 3.1
  termsOfService: 'https://jumkid.com/terms-of-use'
  contact:
    name: Chooli Yip
    url: jumkid.com
    email: chooli.yip@jumkid.com
  license:
    name: Jumkid Public License
    url: 'http://license.jumkid.com'

servers:
  - description: Testing
    url: 'http://test.jumkid.com'
  - description: Live production
    url: 'http://api.jumkid.com'

tags:
  - name: v1
    description: API for version 1.0.0
  - name: test
    description: API for test

security:
  - bearerAuth: []

paths:
  '/activities':
    get:
      tags: ['v1']
      operationId: getActivities
      description: get all activities by user
      parameters:
        - name: entityId
          in: query
          required: false
          description: activity link entity identifier
          schema:
            type: string
        - name: entityName
          in: query
          required: false
          description: activity link entity name
          schema:
            type: string    
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'

    post:
      tags: ['v1']
      operationId: addActivity
      description: add new activity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'

  '/activities/{activityId}':
    get:
      tags: ['v1']
      operationId: getActivity
      description: get existing activity by id
      parameters:
        - name: activityId
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'

    put:
      tags: ['v1']
      operationId: updateActivity
      description: update existing activity
      parameters:
        - name: activityId
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'

    delete:
      tags: ['v1']
      operationId: deleteActivity
      description: delete existsing activity
      parameters:
        - name: activityId
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
      responses:
        '204':
          description: no content
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  errorCode:
                    type: string
                  msg:
                    type: string
                  data:
                    type: string

  '/activity/priorities':
    get:
      tags: ['v1']
      operationId: getPriorities
      description: get all priorities of activity
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Priority'

  '/activity/statuses':
    get:
      tags: ['v1']
      operationId: getStatuses
      description: get all statues of activity
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityStatus'

  '/activity/content':
    post:
      tags: ['v1']
      operationId: add
      description: add new content resource to activity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentResource'
      responses:
        '200':        
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResource'

  '/activity/content/{id}':           
    delete:
      tags: ['v1']
      operationId: delete
      description: delete a content resource
      parameters:
        - name: id
          in: path
          required: true
          description: content identifier
          schema:
            type: string
      responses:
        '204':        
          description: NO_CONTENT

components:
  schemas:
    Activity:
      description: Model of activity
      properties:
        id:
          type: integer
        name:
          required: true
          type: string
        description:
          type: string
        status:
          required: true
          $ref: '#/components/schemas/ActivityStatus'
        priority:
          required: true
          $ref: '#/components/schemas/Priority'
        startDate:
          required: true
          $ref: 'https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime'
        endDate:
          $ref: 'https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime'
        autoNotify:
          type: boolean
        activityNotifications:
          type: array
          items:
            $ref: '#/components/schemas/ActivityNotification'
        ActivityEntityLinks:
          type: array
          items:
            $ref: '#/components/schemas/ActivityEntityLink'    
        createdBy:
          type: string
        creationDate:
          $ref: 'https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime'
        modifiedBy:
          type: string
        modificationDate:
          $ref: 'https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime'

    ActivityStatus:
      type: string
      enum: [DRAFT, IN_PROGRESS, PENDING, COMPLETED]

    NotifyTimeUnit:
      type: string
      enum: [MINUTE, HOUR, DAY]

    ActivityNotification:
      description: model of notification
      properties:
        id:
          type: integer
        expired:
          type: boolean
        notifyBefore:
          type: integer
        notifyBeforeUnit:
          $ref: '#/components/schemas/NotifyTimeUnit'

    ActivityEntityLink:
      description: model of activity entity link
      properties:
        id:
          type: integer
        activityId:
          type: integer
        entityId:
          type: string
        entityName:
          type: string       

    ContentResource:
      description: model of content resource
      properties:
        id:
          type: integer
        activityId:
          type: integer
        contentResourceId:
          type: string

    Priority:
      description: model of priority
      properties:
        id:
          type: integer
        identifier:
          type: string
        label:
          type: string 

    securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT